requirements.txt
openai-whisper
elevenlabs
pyaudio
playsound